+++ pkg/proc/interface.go
@@ -69,8 +69,6 @@ type Info interface {
 	Valid() (bool, error)
 	BinInfo() *BinaryInfo
 	EntryPoint() (uint64, error)
-	// Common returns a struct with fields common to all backends
-	Common() *CommonProcess
 
 	ThreadInfo
 	GoroutineInfo
@@ -94,7 +92,7 @@ type GoroutineInfo interface {
 type ProcessManipulation interface {
 	ContinueOnce() (trapthread Thread, err error)
 	SwitchThread(int) error
-	SwitchGoroutine(int) error
+	SwitchGoroutine(*G) error
 	RequestManualStop() error
 	// CheckAndClearManualStopRequest returns true the first time it's called
 	// after a call to RequestManualStop.
@@ -108,23 +106,4 @@ type BreakpointManipulation interface {
 	SetBreakpoint(addr uint64, kind BreakpointKind, cond ast.Expr) (*Breakpoint, error)
 	ClearBreakpoint(addr uint64) (*Breakpoint, error)
 	ClearInternalBreakpoints() error
-}
-
-// CommonProcess contains fields used by this package, common to all
-// implementations of the Process interface.
-type CommonProcess struct {
-	goroutineCache
-}
-
-type goroutineCache struct {
-	partialGCache map[int]*G
-	allGCache     []*G
-
-	allgentryAddr, allglenAddr uint64
-}
-
-// NewCommonProcess returns a struct with fields common across
-// all process implementations.
-func NewCommonProcess() CommonProcess {
-	return CommonProcess{}
 }
